Criando CRUD de services

+++ backend/src/services/product/DeleteProductService.ts
@@ -1,4 +1,4 @@
-import { DeleteProductError } from "../../exceptions/ProductErrors.js";
+import { DeleteProductError, ProductNotFoundError } from "../../exceptions/ProductErrors.js";
 import prismaClient from "../../prisma/index.js";
 
 interface DeleteProductServiceProps {
@@ -8,6 +8,16 @@ interface DeleteProductServiceProps {
 class DeleteProductService {
     async execute({ product_id }: DeleteProductServiceProps) {
         try {
+            const productExists = await prismaClient.product.findFirst({
+                where: {
+                    id: product_id,
+                },
+            });
+
+            if (!productExists) {
+                throw new ProductNotFoundError();
+            }
+            
             await prismaClient.product.update({
                 where: {
                     id: product_id,
@@ -20,6 +30,9 @@ class DeleteProductService {
             return { message: "Produto deletado/arquivado com sucesso" };
         } catch (error) {
             console.log(error);
+            if(error instanceof ProductNotFoundError){
+                throw error;
+            }
             throw new DeleteProductError();
         }
     }
